Move watcher pkg to the buffer/watcher
